Improved Javadocs

+++ InlineMenuAPI/src/main/java/com/github/unafraid/telegrambot/handlers/inline/InlineContext.java
@@ -29,14 +29,24 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class InlineContext
 {
-	private final Map<Integer, InlineUserData> usersData = new ConcurrentHashMap<>();
+	private final Map<Long, InlineUserData> usersData = new ConcurrentHashMap<>();
 	
-	public InlineUserData getUserData(int id)
+	/**
+	 * Returns user data by user id
+	 * @param id user id
+	 * @return the inline user data
+	 */
+	public InlineUserData getUserData(long id)
 	{
 		return usersData.computeIfAbsent(id, InlineUserData::new);
 	}
 	
-	public boolean clear(int id)
+	/**
+	 * Removes data by user id
+	 * @param id the user id
+	 * @return whether the user data was removed or not
+	 */
+	public boolean clear(long id)
 	{
 		return usersData.remove(id) != null;
 	}
